Addressed review comments Moved CheckpointFileWithFailureHandler to storage.internals.checkpoint and moved LeaderEpochFileCache to storage.internals.epoch
